Refactor: Move function definition
This will alloow reuse the code in the `create_table` function
accross multiple scripts.
