+++ utils/gen_srg_table.py
@@ -9,8 +9,7 @@ import ssg.build_stig
 import ssg.build_yaml
 import ssg.constants
 import ssg.jinja
-from ssg.constants import PREFIX_TO_NS as NS
-from utils.template_renderer import FlexibleLoader
+from utils.gen_tables_common import create_table
 
 
 def parse_args():
@@ -25,16 +24,6 @@ def parse_args():
     parser.add_argument("srgmap", help="Output SRG map path")
     parser.add_argument("srgmap_flat", help="Output flat SRG map path")
     return parser.parse_args()
-
-
-def create_table(data, template_name, output_filename):
-    html_jinja_template = os.path.join(
-        os.path.dirname(__file__), "tables", template_name)
-    env = ssg.jinja._get_jinja_environment(dict())
-    env.loader = FlexibleLoader(os.path.dirname(html_jinja_template))
-    result = ssg.jinja.process_file(html_jinja_template, data)
-    with open(output_filename, "wb") as f:
-        f.write(result.encode('utf8', 'replace'))
 
 
 def get_rules_by_srgid(build_dir, product):
